AI_Agent/multimodal_backend/requirements.txt
# CtrlChecks Multimodal AI Backend Requirements
# Python 3.9+

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# AI/ML Models (Local CPU)
torch>=2.0.0
# transformers>=4.35.0  # Removed - using Ollama instead
sentencepiece>=0.1.99
# diffusers>=0.21.0  # Removed - Ollama doesn't support text-to-image
# accelerate>=0.24.0  # Only needed for transformers/diffusers

# Image Processing
Pillow>=10.0.0

# Optional: For local model loading (if needed in future)
# torch==2.1.0
# transformers==4.35.0
# sentencepiece==0.1.99

# Note: We use HuggingFace Inference API, NOT local models
# This keeps memory usage low and deployment simple



# Validated Cloud API
# huggingface_hub>=0.20.0  # Removed - using Ollama instead

# Audio Processing
soundfile>=0.12.0
librosa>=0.10.0
scipy>=1.10.0
numpy>=1.24.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP client for async requests
aiohttp>=3.9.0
